rag-chatbot-backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database drivers
asyncpg==0.29.0
qdrant-client==1.7.0

# AI/ML
google-generativeai==0.3.2
tiktoken==0.5.2  # For text chunking in MDX parser

# Configuration and environment
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
structlog==24.1.0

# HTTP client for Better Auth
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For testing HTTP endpoints

# Code quality
ruff==0.1.14
mypy==1.8.0
